# Sprint5T2N1_fase1&3/S05T02N01SedoBulpeJordi/S05T02N01SedoBulpeJordi/target/classes/api-docs.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"jugador-controller","description":"REST API relacionada amb les entitats Jugador i Jugada."}],"paths":{"/players/":{"get":{"tags":["jugador-controller"],"summary":"Mostra tots el jugadors","description":"Mostra una llista de jugadors amb els seus percentatges d'èxit. DTO.\n<<Auth.: lectura>>","operationId":"mostrarJugadorsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["jugador-controller"],"summary":"Crea un jugador","description":"Dona d'alta a un nou jugador.\n<<Auth.: tots>>","operationId":"crearJugadorUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"jugador","description":"jugador","required":true,"schema":{"$ref":"#/definitions/Jugador"}}],"responses":{"200":{"description":"amb èxit|OK","schema":{"type":"object"}},"201":{"description":"Created"},"400":{"description":"ja existeix aquest jugador"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["jugador-controller"],"summary":"Modifica els atributs d'un jugador","description":"Canvia els atributs, segons el seu id.\n<<Auth.:escriptura>>","operationId":"modificarJugadorUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"jugador","description":"jugador","required":true,"schema":{"$ref":"#/definitions/Jugador"}}],"responses":{"200":{"description":"amb èxit|OK","schema":{"type":"object"}},"201":{"description":"Created"},"400":{"description":"No es pot actualitzar un jugador que no existeix o bé no es pot modificar el registre de creació!!!"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/players/ranking":{"get":{"tags":["jugador-controller"],"summary":"Mostra el ranking global","description":"Mostra el percentatge d'èxit mig de tots els jugadors.\n<<Auth.: lectura>>","operationId":"mostrarRankingsGlobalsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/players/ranking/loser":{"get":{"tags":["jugador-controller"],"summary":"Mostra el pitjor jugador","description":"Mostra el jugador amb pitjor rendiment.\n<<Auth.: lectura>>","operationId":"mostrarPitjorJugadorUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/players/ranking/winner":{"get":{"tags":["jugador-controller"],"summary":"Mostra el millor jugador","description":"Mostra el jugador amb millor rendiment.\n<<Auth.: lectura>>","operationId":"mostrarMillorJugadorUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/players/{id}/games":{"get":{"tags":["jugador-controller"],"summary":"Mostra les jugades d'un jugador","description":"Mostra una llista amb totes les jugades enregistrades del jugador l'id del qual es passa per paràmetre.\n<<Auth.: lectura>>","operationId":"mostraJugadesJugadorUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["jugador-controller"],"summary":"Fer una tirada de daus","description":"Se li passa l'id d'un jugador i realitza una tirada que s'associarà amb aquest.\n<<Auth.: ROL 'player'>>","operationId":"ferTiradaUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"amb èxit|OK","schema":{"type":"object"}},"201":{"description":"Created"},"400":{"description":"No existeix el jugador!!!"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["jugador-controller"],"summary":"Esborra les jugades d'un jugador","description":"Elimina les jugades d'un jugador de la BBDD.\n<<Auth.: escriptura>>","operationId":"eliminarTiradesJugadorUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"amb èxit|OK","schema":{"type":"object"}},"204":{"description":"No Content"},"400":{"description":"S'ha passat un id que no existeix!!!"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}}},"definitions":{"Jugador":{"type":"object","required":["nom","pk_jugador","registre"],"properties":{"nom":{"type":"string","description":"Nom del jugador"},"pk_jugador":{"type":"integer","format":"int32","description":"Id del jugador - autoincrementat"},"registre":{"type":"string","format":"date-time","description":"Data alta de nou jugador"}},"title":"Jugador"}}}